Adds services views mutliple URL service endpoints.
This commit introduces the ability to utilizing Services Views mutliple service URL endpoints. The two that are currently supported are:
* <domain>/<path-to-endpoint>/<resource-id>
* <domain>/<path-to-endpoint>/views/<view-machine-name> (+optional query string params)

The viewGet method can be called as it had been called before with no regression issues. However, with the introduction of "callbackType", the caller now has the option of utlizing the different services end points that Services Views supports.